refactor: página de anúncio e de gerenciar vendas
