--- src/main/webapp/paginas/admin/gerenciarVendas/gerenciarVendas.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Gerenciar Pedidos</title>
-    <link rel="stylesheet" href="/css/style.css">
-</head>
-
-<body class="bg-light">
-    <nav id="main-nav"></nav>
-
-    <div id="loading" class="container my-5" style="display: flex; height: 70%; width: 100%; align-items: center; justify-content: center;">
-        <img src="/img/loading.gif" style="height: 15%;">
-    </div>
-
-    <div id="main-container" class="container my-5">
-        <h2 class="text-center mb-4">Gestão de Pedidos</h2>
-
-        <!-- Table for Managing Orders -->
-        <table class="table table-bordered table-striped">
-            <thead class="table-dark">
-                <tr>
-                    <th>#</th>
-                    <th>Pedido</th>
-                    <th>Data</th>
-                    <th>CEP</th>
-                    <th>Endereço</th>
-                    <th>Valor total</th>
-                    <th>Frete</th>
-                    <th>Cliente</th>
-                    <th>Status</th>
-                    <th>Operações</th>
-                </tr>
-            </thead>
-            <tbody id="tabela-pedidos"></tbody>
-        </table>
-
-        <h2 class="text-center mb-4">Pedidos de troca e devolução</h2>
-        <table id="secao-pedidos-troca" class="table table-bordered table-striped">
-            <thead class="table-dark">
-                <tr>
-                    <th>#</th>
-                    <th>Pedido</th>
-                    <th>Data</th>
-                    <th>Cliente</th>
-                    <th>Quadrinho</th>
-                    <th>Status</th>
-                    <th>Operações</th>
-                </tr>
-            </thead>
-            <tbody id="tabela-pedido-troca"></tbody>
-        </table>
-    </div>
-
-    <script type="module" src="/js/loadAdminPage.js"></script>
-    <script type="module" src="/paginas/admin/gerenciarVendas/js/gerenciarVendas.js"></script>
-</body>
-
-</html>
