+++ src/main/webapp/paginas/anuncio.html
@@ -1,0 +1,34 @@
+<title>Anúncio</title>
+
+<nav id="main-nav"></nav>
+
+<div id="container-anuncio" class="container my-5 flex-grow-1">
+  <div class="card shadow-lg">
+    <div class="row g-0">
+
+      <section id="secao-imagem" class="col-md-6"></section>
+
+      <section id="secao-principal" class="col-md-6">
+        <div class="card-body">
+          <section id="header-produto" class="d-flex flex-column align-items-center gap-3"></section>
+
+          <hr class="my-4">
+
+          <section id="adicionar-carrinho" class="d-flex flex-column align-items-center gap-3">
+            <div class="d-flex gap-3 w-75">
+              <button class="btn btn-warning btn-lg" onclick="adicionarItem()">Adicionar ao carrinho</button>
+              <input type="number" name="quantidade" class="form-control" min="1" placeholder="Quantidade" value=1>
+            </div>
+            <p id="estoque"></p>
+          </section>
+          <hr class="my-4">
+
+          <h3 class="mt-4">Detalhes</h3>
+          <table id="tabela-info"class="table table-hover"></table>
+        </div>
+      </section>
+    </div>
+  </div>
+</div>
+
+<footer id="main-footer" class="text-black text-center py-4"></footer>
