--- src/main/webapp/paginas/cliente/anuncio/anuncio.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Anúncio</title>
-  <link rel="stylesheet" href="/css/style.css">
-</head>
-
-<body class="bg-light d-flex flex-column">
-
-  <nav id="main-nav"></nav>
-
-  <div id="container-anuncio" class="container my-5 flex-grow-1">
-    <div class="card shadow-lg">
-      <div class="row g-0">
-
-        <section id="secao-imagem" class="col-md-6"></section>
-
-        <section id="secao-principal" class="col-md-6">
-          <div class="card-body">
-            <section id="header-produto" class="d-flex flex-column align-items-center gap-3"></section>
-
-            <hr class="my-4">
-
-            <section id="adicionar-carrinho" class="d-flex flex-column align-items-center gap-3">
-              <div class="d-flex gap-3 w-75">
-                <button class="btn btn-warning btn-lg" onclick="adicionarItem()">Adicionar ao carrinho</button>
-                <input type="number" name="quantidade" class="form-control" min="1" placeholder="Quantidade" value=1>
-              </div>
-              <p id="estoque"></p>
-            </section>
-            <hr class="my-4">
-
-            <h3 class="mt-4">Detalhes</h3>
-            <table id="tabela-info"class="table table-hover"></table>
-          </div>
-        </section>
-      </div>
-    </div>
-  </div>
-
-  <footer id="main-footer" class="text-black text-center py-4"></footer>
-
-  <script type="module" src="/js/loadPage.js"></script>
-  <script type="module" src="/paginas/cliente/anuncio/js/anuncio.js"></script>
-</body>
